fix: 자주 묻는 질문 api 연결 수정
